# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  legocy-go_internal_delievery_http_resources_users.UserImageInfoResponse:
    properties:
      downloadURL:
        type: string
      filepath:
        type: string
      userID:
        type: integer
    type: object
  legocy-go_internal_delievery_http_resources_users.UserImageUploadResponse:
    properties:
      imageURL:
        type: string
      ok:
        type: boolean
    type: object
  legocy-go_internal_delievery_http_resources_users.UserImagesListResponse:
    properties:
      images:
        items:
          $ref: '#/definitions/legocy-go_internal_delievery_http_resources_users.UserImageInfoResponse'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: LEGOcy is a marketplace for LEGO lovers.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: LEGOcy API
  version: "1.0"
paths:
...
  /users/images/:userID:
    get:
      operationId: get_user_images
      parameters:
      - description: user ID
        in: path
        name: userID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/legocy-go_internal_delievery_http_resources_users.UserImagesListResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Get User Images List
      tags:
      - users_images
    post:
      consumes:
      - multipart/form-data
      operationId: upload_user_image
      parameters:
      - description: filepath
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/legocy-go_internal_delievery_http_resources_users.UserImageUploadResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Download User Image
      tags:
      - users_images
...
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"